Added: LogMessageBase: abstract class for log messages. MessageLog: storage and provider of log messages. EventMessageProvider: static container similar to CombatProvider but for non-combat log messages. Also made messages able to generate their messages from const strings and a number of saved variables, instead of saving the entire text string each time.
Input:
Added command type key-binds for non-gameplay functionality, such as ctrl-s for saving the game, and keys for scrolling the log (pgdn, pgup).
Made GameLoop implement IInputEndpoint to handle the save command.

Rendering:
Major rewrite of the log to handle messages as objects with a scrollable history. Also fixed both the status bar and the log to not line-break mid word.

Player:
The overriding properties Name and Health of PlayerEntity interfered with MongoDB serialization so they were renamed and made private.

Level entities:
+BsonConstructors
Implemented use of LogMessage

Level:
Fixed a bug where a few tiles were incorrectly marked as obscured when near the bottom layer of the map.

MongoDB:
MongoDB Driver can now automatically serialize everything needed to save the game.
Refined the save functionality and added the MessageLog to the save.

Also cleaned up a lot of usings references