Added log messages to the saving, storing it in a separate MongoDB collection "GameLogs". The idea is to let the game log be available after the character has died for review, and to get a preview of the log in the load menu before selecting which save to load.
Added a load save method to GameMongoClient.
Program calls this at startup and loads the first available save.
And a load log method which is used by load save and will be used to preview saves in the menu.

Added NameHistory class which can be used to keep track of and prevent usage of the same name several times.

Changed the way InputHandler and GameMongoClient Instance singleton is controlled.
Fixed Ctrl+S, the keyChar returned from the keystroke is not 'S', it's \u0013 which is ctrl.

Fixed a number of small log bugs